feat: extract List methods into separated file
