+++ packages/app-folders/src/components/Tree/List/constants.ts
@@ -1,0 +1,1 @@
+export const ROOT_ID = "root";
